editoast: move supported signaling systems to shcemas
