+++ editoast/src/schema/rolling_stock/light_rolling_stock.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 
+use editoast_schemas::rolling_stock::RollingStockSupportedSignalingSystems;
 use serde::Deserialize;
 use serde::Serialize;
 use utoipa::ToSchema;
@@ -10,7 +11,6 @@ use super::Gamma;
 use super::RollingResistance;
 use super::RollingStockMetadata;
 use crate::modelsv2::rolling_stock_livery::RollingStockLiveryMetadataModel;
-use crate::modelsv2::rolling_stock_model::RollingStockSupportedSignalingSystems;
 use crate::schema::track_section::LoadingGaugeType;
 
 editoast_common::schemas! {
